+++ instrumentation/github.com/labstack/echo/otelecho/example/server.go
@@ -40,7 +40,7 @@ func main() {
 		}
 	}()
 	r := echo.New()
-	r.Use(otelecho.Middleware("my-server", otelecho.WithSpanNameFormatter(otelecho.PathSpanNameFormatter)))
+	r.Use(otelecho.Middleware("my-server", otelecho.WithSpanNameFormatter(otelecho.PathSpanNameFormatter), otelecho.WithRouteTagFromPath()))
 
 	r.GET("/users/:id", func(c echo.Context) error {
 		id := c.Param("id")