+++ instrumentation/go.mongodb.org/mongo-driver/mongo/otelmongo/mongo.go
@@ -108,7 +108,7 @@ func (m *monitor) Finished(evt *event.CommandFinishedEvent, err error) {
 }
 
 // TODO sanitize values where possible
-// TODO limit maximum size
+// TODO limit maximum size.
 func sanitizeCommand(command bson.Raw) string {
 	b, _ := bson.MarshalExtJSON(command, false, false)
 	return string(b)